Add explanatory code-comment
In TPreloadedFile.Open, if the file does not yet exist we create it
and then immediately close it, only to open it again right afterwards.

This is necessary as Windows would ignore fmShareDenyNone if fmCreate
is also used. So we need to *always* open an already existing file
with fmShareDenyNone.