Bugfix: GenDb incorrectly calculates ranking-data
This mainly caused the backlink-data to be incorrect, but
other ranking-data as well. Though the backlink-data had
the most severe effect on ranking-quality.

This fix also caused significant performance-improvement
in index generation. This is because two previously separate
passes are now integrated into the first pass.

This bug was introduced with recent changes for increasing
maximum index-size. These changed into which index-shard
URL-data was put, and thus caused the later index-passes
to look in the wrong places for ranking-data.