Enable SearchServerNew to run many instances
While SearchServerNew could be run in multiple instances before,
there was no easy way to do this.

These changes allow the path of the index to be passed as a commandline-
parameter instead of having to have a config-file for each instance.

Also fixed a bug which caused much more memory to be allocated for
some cached data than necessary. For each page in the index 12 bytes
more than necessary were allocated, meaning that for an index with
1 billion pages there would have been 12gb more RAM-usage than needed.