Implement index-compression in search-server
Index-compression is only implemented in searchservernew, NOT
in the old SearchServer!

There is a feature-flag that at runtime determines if
index-compression is used. This is in a text-file named
features.cfg, which is located in the directory of the index.

If features.cfg contains a line "EnableCompressedIndex=true", then
this feature is enabled. The old-style keysXX.idx files are NOT
used in that case - they don't even need to be there - instead the
new-style keysXX.cidx files are used.

Index-compression cuts down the index-files to just a little more
than half their original size.