Reduce memory consumption of searchservernew
BackLinkCounts went from int64 to int32 in memory.

Note that the on-disk data is still int64. Data is loaded as int64
and then copied to int32. Conversion makes sure that MAXINT is not
exceeded.

There is also code added to include the inverse-document-frequency
as a ranking-factor. This is still considered experimental, although
it seems to be working fine.