Change implementation of ValueData-array
The ValueData-array is used for a bucket-sort of search-results.

The implementation previously used a fixed-size two-dimensional array.
This resulted in huge memory usage (8gb while running 32 instances of
searchservernew).

The new implementation uses a fixed-size one-dimensional array
containing variable-size arrays as its elements. Memory usage of
ValueData is now less than 1/10 of what it was before.

This change became necessary when having an index of 1 billion pages,
split up into 32 sub-indeces, used up almost of the memory on a server
with 32gb RAM. Now there is breathing-room again. :)