searchservernew: Re-arrange data for performance
A lot of internal data-structures in searchservernew have been
re-arranged for much better performance.

Previously the ordering of internal data meant that the CPU-cache had
LOTS of cache-misses. These changes fix that. Queries are almost twice
as fast now.

On the downside, startup is now considerably slower, as data has to be
re-arranged while being loaded into RAM, causing LOTS of cache-misses
at that time.

With a 1-billion-pages index split into 32 parts (meaning 32 instances
of searchservernew), startup time went from 10 seconds to about
40 seconds because about 10gb of data has to be copied around in
RAM in an order that is really disadvantageous for the CPU-cache. The
performance-gain during queries is worth this though.