Fix broken "host:xx.tld" in searchservernew.dpr
Found this bug during some major code-cleanup. So this commit
also covers that code-cleanup.

Please note that currently searchservernew won't work correctly
on Linux due to a bug in the Indy library that I haven't found a
workaround or fix for yet. The problem seems to be caused by
incorrect conversion between code-tables (ASCII <-> UTF8).
This causes texts like "schöbel" to be broken at the umlaut ("sch").
This only happens on Linux. Compiling for Windows with FreePascal
works fine.