Add compressrwi.dpr
This little commandline tool is used to compress the
Reverse-Word-Index created by GenDb. Saves a lot of
disk-space at the sacrifice of a little bit of performance.
Though in some instances it can even improve performance
as less data has to be read from disk during a search.

It existed for a few months now, but I had simply forgot to
commit it.

During development I had added it to .gitignore which made
it easy to forget.