Temporary bugfix for Indy on Linux
This fixes the problem that query-parameters and HTTP-server
response-content which contain Non-ASCII characters will not work
on Linux.

This bug affected searchservernew.

I'm not sure if this fixes the *actual* problem with Indy. Especially
I don't know what this will do to MacOS. But after the fix both Windows
and Linux work fine. And that's all that *I* care about. :)