Add option -DeOnly to PrepareRobot
Actually this ended up working as "-DeAtChOnly".

If used as a commandline-option, PrepareRobot will only output URLs
on .de, .at, and .ch domains. These are (mostly) in German, so I
kept the actual option's name as "-DeOnly".

I implemented this to have a way to increase the percentage of
German-language pages in the index.