Rework shard-selection
This commit makes rather big changes to the way shards are selected
in GenDb, ImportUrls and PrepareRobot.

While this was previously done with just starting and ending shard
numbers, it now adds a shard-mask. Starting and ending numbers
refer to the shard-number after it has been AND-ed with the shard
mask.

This was necessary to ease the switch from 128 to 512 shards that
are now used. Without the shard-mask it would have been a pain in
the b*** to handle this.

This commit also includes major changes to the CLI parameters for
GenDb, ImportUrls and PrepareRobot.